requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
protobuf==5.28.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
PyJWT==2.9.0
cryptography==43.0.1
